+++ src/main/java/roomescape/member/controller/AdminPageController.java
@@ -1,32 +1,34 @@
-package roomescape.admin;
+package roomescape.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/admin")
 public class AdminPageController {
 
-    @GetMapping("/admin")
+    @GetMapping
     public String adminPage() {
         return "admin/index";
     }
 
-    @GetMapping("/admin/reservation")
+    @GetMapping("/reservation")
     public String adminReservationPage() {
         return "admin/reservation-new";
     }
 
-    @GetMapping("/admin/time")
+    @GetMapping("/time")
     public String timePage() {
         return "admin/time";
     }
 
-    @GetMapping("/admin/theme")
+    @GetMapping("/theme")
     public String adminThemePage() {
         return "admin/theme";
     }
 
-    @GetMapping("/admin/waiting")
+    @GetMapping("/waiting")
     public String adminWaitingManagePage() {
         return "admin/waiting";
     }
