+++ src/main/java/roomescape/member/controller/UserPageController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class MemberPageController {
+public class UserPageController {
 
     @GetMapping("/signup")
     public String userSignUpPage() {
