+++ src/test/java/roomescape/member/controller/AdminPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.member.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
